+++ pyhapi/__init__.py
@@ -9,9 +9,10 @@ from .hgeo import HGeo, HGeoMesh, HGeoCurve, HGeoHeightfield, HGeoInstancer
 from .hsession import HSession, HSessionManager, HSessionPool, HSessionTask
 from .hnode import HNode, HInputNode, HHeightfieldInputNode, HHeightfieldInputVolumeNode
 from .hasset import HAsset
+from .hparm import *
 from . import hapi as HAPI
 
-__version__ = "0.0.3b1"
+__version__ = "0.0.4b0"
 
 
 import platform